Amended the README and fixed a bug in the API.

+++ client/src/main/kotlin/net/corda/examples/oracle/api/ClientApi.kt
@@ -53,12 +53,12 @@ class ClientApi(val services: CordaRPCOps) {
     }
 
     /**
-     * Creates a new priume number by consulting the primes Oracle.
+     * Creates a new prime number by consulting the primes Oracle.
      */
     @GET
     @Path("create-prime")
     @Produces(MediaType.APPLICATION_JSON)
-    fun createPrime(@QueryParam(value = "n") n: Int): Response {
+    fun createPrime(@QueryParam(value = "n") n: Long): Response {
         // Start the CretePrime flow. We block and wait for the flow to return.
         val (status, message) = try {
             val flowHandle = services.startFlowDynamic(CreatePrime::class.java, n)
